requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
pandas
jupyter
opencv-python
pillow
seaborn
gradio